@@ -1,226 +1,455 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <UserSecretsId>User-Secret-ID</UserSecretsId>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <UserSecretsId>0e7b0af0-23c1-4720-acb5-ccf5ffc15a8e</UserSecretsId>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
-
-  <PropertyGroup>
-    <UserSecretsId>aspnet-RenewalWebsite-d3947413-e3ea-45b0-8884-7a058d6ff7b9</UserSecretsId>
-  </PropertyGroup>
-
-
   <ItemGroup>
-    <Compile Remove="Resources\Views.Error.404.zh-CN.Designer.cs" />
-    <Compile Remove="Resources\Views.Error.500.zh-CN.Designer.cs" />
-    <Compile Remove="Resources\Views.Error.DataNotFound.zh-CN.Designer.cs" />
+    <None Include="Views\Home\Index.cshtml" />
+    <None Include="Views\Home\Privacy.cshtml" />
+    <None Include="Views\Shared\Error.cshtml" />
+    <None Include="Views\Shared\_Layout.cshtml" />
+    <None Include="Views\Shared\_ValidationScriptsPartial.cshtml" />
+    <None Include="Views\_ViewImports.cshtml" />
+    <None Include="Views\_ViewStart.cshtml" />
+    <None Include="wwwroot\downloads\renewal-logo.zip" />
+    <None Include="wwwroot\downloads\日新中心简介 Let%27s Renewal 2020-04-28.pdf" />
+    <None Include="wwwroot\fonts\dpt35-webfont.svg" />
+    <None Include="wwwroot\fonts\dpt35-webfont.woff2" />
+    <None Include="wwwroot\fonts\dpt65-webfont.svg" />
+    <None Include="wwwroot\fonts\dpt65-webfont.woff2" />
+    <None Include="wwwroot\fonts\fontawesome-webfont.svg" />
+    <None Include="wwwroot\fonts\fontawesome-webfont.woff2" />
+    <None Include="wwwroot\InvoiceHistory\375abb18-1336-4ef4-8b67-7ce708fb2bf3\Renewal_2018-09-20.pdf" />
+    <None Include="wwwroot\InvoiceHistory\375abb18-1336-4ef4-8b67-7ce708fb2bf3\Renewal_2018-09-21.pdf" />
+    <None Include="wwwroot\InvoiceHistory\375abb18-1336-4ef4-8b67-7ce708fb2bf3\Rixin_2018-09-20.pdf" />
+    <None Include="wwwroot\InvoiceHistory\48ba2f06-e22c-41a8-95a0-0b8eedaf5bf3\Renewal_2018-09-20.pdf" />
+    <None Include="wwwroot\js\cbpFWTabs.js" />
+    <None Include="wwwroot\js\ccFormat.js" />
+    <None Include="wwwroot\js\DatePicker\jquery.datetimepicker.full.js" />
+    <None Include="wwwroot\js\daterangepicker\daterangepicker.js" />
+    <None Include="wwwroot\js\developer.js" />
+    <None Include="wwwroot\js\jconfirm\jquery-confirm.js" />
+    <None Include="wwwroot\js\jconfirm\jquery-confirm.min.js" />
+    <None Include="wwwroot\js\jquery-ui.min.js" />
+    <None Include="wwwroot\js\jquery.buttonLoader.js" />
+    <None Include="wwwroot\js\jquery.buttonLoader.min.js" />
+    <None Include="wwwroot\js\jquery.creditCardValidator.js" />
+    <None Include="wwwroot\js\jquery.inputmask.bundle.js" />
+    <None Include="wwwroot\js\jquery.inputmask.bundle.min.js" />
+    <None Include="wwwroot\js\jquery.scrolling-tabs.js" />
+    <None Include="wwwroot\js\jquery.scrolling-tabs.min.js" />
+    <None Include="wwwroot\js\jquery.unobtrusive-ajax.min.js" />
+    <None Include="wwwroot\js\jquery.validate.min.js" />
+    <None Include="wwwroot\js\jquery.validate.unobtrusive.min.js" />
+    <None Include="wwwroot\js\modernizr.custom.js" />
+    <None Include="wwwroot\js\moment-timezone.js" />
+    <None Include="wwwroot\js\moment.js" />
+    <None Include="wwwroot\js\site.js" />
+    <None Include="wwwroot\js\site.min.js" />
+    <None Include="wwwroot\lib\bootstrap\CHANGELOG.md" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap-grid.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap-grid.min.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap-reboot.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap-reboot.min.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap.min.css.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.bundle.js" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.bundle.js.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.bundle.min.js" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.bundle.min.js.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.js" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.js.map" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.min.js" />
+    <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.min.js.map" />
+    <None Include="wwwroot\lib\bootstrap\Gemfile" />
+    <None Include="wwwroot\lib\bootstrap\Gemfile.lock" />
+    <None Include="wwwroot\lib\bootstrap\Gruntfile.js" />
+    <None Include="wwwroot\lib\bootstrap\grunt\change-version.js" />
+    <None Include="wwwroot\lib\bootstrap\grunt\gcp-key.json.enc" />
+    <None Include="wwwroot\lib\bootstrap\grunt\postcss.js" />
+    <None Include="wwwroot\lib\bootstrap\grunt\sauce_browsers.yml" />
+    <None Include="wwwroot\lib\bootstrap\grunt\upload-preview.sh" />
+    <None Include="wwwroot\lib\bootstrap\ISSUE_TEMPLATE.md" />
+    <None Include="wwwroot\lib\bootstrap\js\.babelrc" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\alert.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\alert.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\button.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\button.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\carousel.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\carousel.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\collapse.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\collapse.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\dropdown.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\dropdown.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\modal.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\modal.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\popover.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\popover.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\scrollspy.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\scrollspy.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\tab.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\tab.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\tooltip.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\tooltip.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\util.js" />
+    <None Include="wwwroot\lib\bootstrap\js\dist\util.js.map" />
+    <None Include="wwwroot\lib\bootstrap\js\src\alert.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\button.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\carousel.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\collapse.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\dropdown.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\modal.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\popover.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\scrollspy.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\tab.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\tooltip.js" />
+    <None Include="wwwroot\lib\bootstrap\js\src\util.js" />
+    <None Include="wwwroot\lib\bootstrap\LICENSE" />
+    <None Include="wwwroot\lib\bootstrap\nuget\bootstrap.nuspec" />
+    <None Include="wwwroot\lib\bootstrap\nuget\bootstrap.sass.nuspec" />
+    <None Include="wwwroot\lib\bootstrap\nuget\MyGet.ps1" />
+    <None Include="wwwroot\lib\bootstrap\package.js" />
+    <None Include="wwwroot\lib\bootstrap\README.md" />
+    <None Include="wwwroot\lib\bootstrap\scss\.scss-lint.yml" />
+    <None Include="wwwroot\lib\bootstrap\scss\bootstrap-grid.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\bootstrap-reboot.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\bootstrap.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_alert.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_background-variant.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_badge.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_border-radius.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_breakpoints.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_buttons.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_cards.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_clearfix.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_float.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_forms.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_gradients.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_grid-framework.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_grid.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_hover.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_image.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_list-group.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_lists.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_nav-divider.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_navbar-align.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_pagination.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_reset-text.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_resize.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_screen-reader.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_size.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_table-row.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_text-emphasis.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_text-hide.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_text-truncate.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_transforms.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\mixins\_visibility.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_align.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_background.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_borders.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_clearfix.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_display.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_flex.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_float.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_position.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_screenreaders.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_sizing.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_spacing.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_text.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\utilities\_visibility.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_alert.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_badge.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_breadcrumb.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_button-group.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_buttons.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_card.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_carousel.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_close.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_code.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_custom-forms.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_custom.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_dropdown.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_forms.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_grid.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_images.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_input-group.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_jumbotron.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_list-group.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_media.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_mixins.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_modal.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_nav.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_navbar.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_normalize.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_pagination.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_popover.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_print.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_progress.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_reboot.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_responsive-embed.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_tables.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_tooltip.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_transitions.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_type.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_utilities.scss" />
+    <None Include="wwwroot\lib\bootstrap\scss\_variables.scss" />
+    <None Include="wwwroot\lib\font-awesome\.gitignore" />
+    <None Include="wwwroot\lib\font-awesome\.npmignore" />
+    <None Include="wwwroot\lib\font-awesome\css\font-awesome.css.map" />
+    <None Include="wwwroot\lib\font-awesome\fonts\fontawesome-webfont.svg" />
+    <None Include="wwwroot\lib\font-awesome\fonts\fontawesome-webfont.woff2" />
+    <None Include="wwwroot\lib\font-awesome\less\animated.less" />
+    <None Include="wwwroot\lib\font-awesome\less\bordered-pulled.less" />
+    <None Include="wwwroot\lib\font-awesome\less\core.less" />
+    <None Include="wwwroot\lib\font-awesome\less\fixed-width.less" />
+    <None Include="wwwroot\lib\font-awesome\less\font-awesome.less" />
+    <None Include="wwwroot\lib\font-awesome\less\icons.less" />
+    <None Include="wwwroot\lib\font-awesome\less\larger.less" />
+    <None Include="wwwroot\lib\font-awesome\less\list.less" />
+    <None Include="wwwroot\lib\font-awesome\less\mixins.less" />
+    <None Include="wwwroot\lib\font-awesome\less\path.less" />
+    <None Include="wwwroot\lib\font-awesome\less\rotated-flipped.less" />
+    <None Include="wwwroot\lib\font-awesome\less\screen-reader.less" />
+    <None Include="wwwroot\lib\font-awesome\less\stacked.less" />
+    <None Include="wwwroot\lib\font-awesome\less\variables.less" />
+    <None Include="wwwroot\lib\font-awesome\scss\font-awesome.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_animated.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_bordered-pulled.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_core.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_fixed-width.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_icons.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_larger.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_list.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_mixins.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_path.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_rotated-flipped.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_screen-reader.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_stacked.scss" />
+    <None Include="wwwroot\lib\font-awesome\scss\_variables.scss" />
+    <None Include="wwwroot\lib\jquery-ajax-unobtrusive\jquery.unobtrusive-ajax.min.js" />
+    <None Include="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.js" />
+    <None Include="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js" />
+    <None Include="wwwroot\lib\jquery-validation\dist\additional-methods.js" />
+    <None Include="wwwroot\lib\jquery-validation\dist\additional-methods.min.js" />
+    <None Include="wwwroot\lib\jquery-validation\dist\jquery.validate.js" />
+    <None Include="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js" />
+    <None Include="wwwroot\lib\jquery-validation\LICENSE.md" />
+    <None Include="wwwroot\lib\jquery\dist\core.js" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.js" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.min.js" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.min.map" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.slim.js" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.slim.min.js" />
+    <None Include="wwwroot\lib\jquery\dist\jquery.slim.min.map" />
+    <None Include="wwwroot\lib\jquery\external\sizzle\dist\sizzle.js" />
+    <None Include="wwwroot\lib\jquery\external\sizzle\dist\sizzle.min.js" />
+    <None Include="wwwroot\lib\jquery\external\sizzle\dist\sizzle.min.map" />
+    <None Include="wwwroot\lib\jquery\README.md" />
+    <None Include="wwwroot\lib\jquery\src\ajax.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\jsonp.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\load.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\parseXML.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\script.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\var\location.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\var\nonce.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\var\rquery.js" />
+    <None Include="wwwroot\lib\jquery\src\ajax\xhr.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes\attr.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes\classes.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes\prop.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes\support.js" />
+    <None Include="wwwroot\lib\jquery\src\attributes\val.js" />
+    <None Include="wwwroot\lib\jquery\src\callbacks.js" />
+    <None Include="wwwroot\lib\jquery\src\core.js" />
+    <None Include="wwwroot\lib\jquery\src\core\access.js" />
+    <None Include="wwwroot\lib\jquery\src\core\DOMEval.js" />
+    <None Include="wwwroot\lib\jquery\src\core\init.js" />
+    <None Include="wwwroot\lib\jquery\src\core\nodeName.js" />
+    <None Include="wwwroot\lib\jquery\src\core\parseHTML.js" />
+    <None Include="wwwroot\lib\jquery\src\core\ready-no-deferred.js" />
+    <None Include="wwwroot\lib\jquery\src\core\ready.js" />
+    <None Include="wwwroot\lib\jquery\src\core\readyException.js" />
+    <None Include="wwwroot\lib\jquery\src\core\stripAndCollapse.js" />
+    <None Include="wwwroot\lib\jquery\src\core\support.js" />
+    <None Include="wwwroot\lib\jquery\src\core\var\rsingleTag.js" />
+    <None Include="wwwroot\lib\jquery\src\css.js" />
+    <None Include="wwwroot\lib\jquery\src\css\addGetHookIf.js" />
+    <None Include="wwwroot\lib\jquery\src\css\adjustCSS.js" />
+    <None Include="wwwroot\lib\jquery\src\css\curCSS.js" />
+    <None Include="wwwroot\lib\jquery\src\css\hiddenVisibleSelectors.js" />
+    <None Include="wwwroot\lib\jquery\src\css\showHide.js" />
+    <None Include="wwwroot\lib\jquery\src\css\support.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\cssExpand.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\getStyles.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\isHiddenWithinTree.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\rmargin.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\rnumnonpx.js" />
+    <None Include="wwwroot\lib\jquery\src\css\var\swap.js" />
+    <None Include="wwwroot\lib\jquery\src\data.js" />
+    <None Include="wwwroot\lib\jquery\src\data\Data.js" />
+    <None Include="wwwroot\lib\jquery\src\data\var\acceptData.js" />
+    <None Include="wwwroot\lib\jquery\src\data\var\dataPriv.js" />
+    <None Include="wwwroot\lib\jquery\src\data\var\dataUser.js" />
+    <None Include="wwwroot\lib\jquery\src\deferred.js" />
+    <None Include="wwwroot\lib\jquery\src\deferred\exceptionHook.js" />
+    <None Include="wwwroot\lib\jquery\src\deprecated.js" />
+    <None Include="wwwroot\lib\jquery\src\dimensions.js" />
+    <None Include="wwwroot\lib\jquery\src\effects.js" />
+    <None Include="wwwroot\lib\jquery\src\effects\animatedSelector.js" />
+    <None Include="wwwroot\lib\jquery\src\effects\Tween.js" />
+    <None Include="wwwroot\lib\jquery\src\event.js" />
+    <None Include="wwwroot\lib\jquery\src\event\ajax.js" />
+    <None Include="wwwroot\lib\jquery\src\event\alias.js" />
+    <None Include="wwwroot\lib\jquery\src\event\focusin.js" />
+    <None Include="wwwroot\lib\jquery\src\event\support.js" />
+    <None Include="wwwroot\lib\jquery\src\event\trigger.js" />
+    <None Include="wwwroot\lib\jquery\src\exports\amd.js" />
+    <None Include="wwwroot\lib\jquery\src\exports\global.js" />
+    <None Include="wwwroot\lib\jquery\src\jquery.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\buildFragment.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\getAll.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\setGlobalEval.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\support.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\var\rcheckableType.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\var\rscriptType.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\var\rtagName.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\wrapMap.js" />
+    <None Include="wwwroot\lib\jquery\src\manipulation\_evalUrl.js" />
+    <None Include="wwwroot\lib\jquery\src\offset.js" />
+    <None Include="wwwroot\lib\jquery\src\queue.js" />
+    <None Include="wwwroot\lib\jquery\src\queue\delay.js" />
+    <None Include="wwwroot\lib\jquery\src\selector-native.js" />
+    <None Include="wwwroot\lib\jquery\src\selector-sizzle.js" />
+    <None Include="wwwroot\lib\jquery\src\selector.js" />
+    <None Include="wwwroot\lib\jquery\src\serialize.js" />
+    <None Include="wwwroot\lib\jquery\src\traversing.js" />
+    <None Include="wwwroot\lib\jquery\src\traversing\findFilter.js" />
+    <None Include="wwwroot\lib\jquery\src\traversing\var\dir.js" />
+    <None Include="wwwroot\lib\jquery\src\traversing\var\rneedsContext.js" />
+    <None Include="wwwroot\lib\jquery\src\traversing\var\siblings.js" />
+    <None Include="wwwroot\lib\jquery\src\var\arr.js" />
+    <None Include="wwwroot\lib\jquery\src\var\class2type.js" />
+    <None Include="wwwroot\lib\jquery\src\var\concat.js" />
+    <None Include="wwwroot\lib\jquery\src\var\document.js" />
+    <None Include="wwwroot\lib\jquery\src\var\documentElement.js" />
+    <None Include="wwwroot\lib\jquery\src\var\fnToString.js" />
+    <None Include="wwwroot\lib\jquery\src\var\getProto.js" />
+    <None Include="wwwroot\lib\jquery\src\var\hasOwn.js" />
+    <None Include="wwwroot\lib\jquery\src\var\indexOf.js" />
+    <None Include="wwwroot\lib\jquery\src\var\ObjectFunctionString.js" />
+    <None Include="wwwroot\lib\jquery\src\var\pnum.js" />
+    <None Include="wwwroot\lib\jquery\src\var\push.js" />
+    <None Include="wwwroot\lib\jquery\src\var\rcssNum.js" />
+    <None Include="wwwroot\lib\jquery\src\var\rnothtmlwhite.js" />
+    <None Include="wwwroot\lib\jquery\src\var\slice.js" />
+    <None Include="wwwroot\lib\jquery\src\var\support.js" />
+    <None Include="wwwroot\lib\jquery\src\var\toString.js" />
+    <None Include="wwwroot\lib\jquery\src\wrap.js" />
+    <None Include="wwwroot\lib\selectbox\fastclick.js" />
+    <None Include="wwwroot\lib\selectbox\jquery.nice-select.js" />
+    <None Include="wwwroot\lib\selectbox\jquery.nice-select.min.js" />
+    <None Include="wwwroot\lib\selectbox\prism.js" />
+    <None Include="wwwroot\lib\tether\CHANGELOG.md" />
+    <None Include="wwwroot\lib\tether\CONTRIBUTING.md" />
+    <None Include="wwwroot\lib\tether\dist\js\tether.js" />
+    <None Include="wwwroot\lib\tether\dist\js\tether.min.js" />
+    <None Include="wwwroot\lib\tether\docs\1-Overview\1-why_you_should_use_tether.md" />
+    <None Include="wwwroot\lib\tether\docs\1-Overview\2-repositioning.md" />
+    <None Include="wwwroot\lib\tether\docs\1-Overview\3-why_we_dont_support_IE_8.md" />
+    <None Include="wwwroot\lib\tether\docs\2-Examples\1-list_of_examples.md" />
+    <None Include="wwwroot\lib\tether\docs\2-Examples\2-projects_using_tether.md" />
+    <None Include="wwwroot\lib\tether\docs\3-Advanced\1-embedding_tether.md" />
+    <None Include="wwwroot\lib\tether\docs\3-Advanced\2-extending_tether.md" />
+    <None Include="wwwroot\lib\tether\docs\coffee\intro.coffee" />
+    <None Include="wwwroot\lib\tether\docs\intro.md" />
+    <None Include="wwwroot\lib\tether\docs\js\intro.js" />
+    <None Include="wwwroot\lib\tether\docs\js\markAttachment.js" />
+    <None Include="wwwroot\lib\tether\docs\sass\intro.sass" />
+    <None Include="wwwroot\lib\tether\docs\welcome\coffee\welcome.coffee" />
+    <None Include="wwwroot\lib\tether\docs\welcome\js\drop.js" />
+    <None Include="wwwroot\lib\tether\docs\welcome\js\jquery.js" />
+    <None Include="wwwroot\lib\tether\docs\welcome\js\log.js" />
+    <None Include="wwwroot\lib\tether\docs\welcome\js\tether-v0.1.3.js" />
+    <None Include="wwwroot\lib\tether\docs\welcome\js\welcome.js" />
+    <None Include="wwwroot\lib\tether\docs\welcome\sass\browser-demo.sass" />
+    <None Include="wwwroot\lib\tether\docs\welcome\sass\welcome.sass" />
+    <None Include="wwwroot\lib\tether\docs\welcome\sass\_inline-block.sass" />
+    <None Include="wwwroot\lib\tether\examples\chosen\chosen.js" />
+    <None Include="wwwroot\lib\tether\examples\dolls\dolls.js" />
+    <None Include="wwwroot\lib\tether\examples\resources\js\jquery.js" />
+    <None Include="wwwroot\lib\tether\examples\resources\js\log.js" />
+    <None Include="wwwroot\lib\tether\gulpfile.js" />
+    <None Include="wwwroot\lib\tether\LICENSE" />
+    <None Include="wwwroot\lib\tether\README.md" />
+    <None Include="wwwroot\lib\tether\src\css\helpers\_tether-theme-arrows.sass" />
+    <None Include="wwwroot\lib\tether\src\css\helpers\_tether-theme-basic.sass" />
+    <None Include="wwwroot\lib\tether\src\css\helpers\_tether.sass" />
+    <None Include="wwwroot\lib\tether\src\css\mixins\_inline-block.sass" />
+    <None Include="wwwroot\lib\tether\src\css\mixins\_pie-clearfix.sass" />
+    <None Include="wwwroot\lib\tether\src\css\tether-theme-arrows-dark.sass" />
+    <None Include="wwwroot\lib\tether\src\css\tether-theme-arrows.sass" />
+    <None Include="wwwroot\lib\tether\src\css\tether-theme-basic.sass" />
+    <None Include="wwwroot\lib\tether\src\css\tether.sass" />
+    <None Include="wwwroot\lib\tether\src\js\abutment.js" />
+    <None Include="wwwroot\lib\tether\src\js\constraint.js" />
+    <None Include="wwwroot\lib\tether\src\js\markAttachment.js" />
+    <None Include="wwwroot\lib\tether\src\js\shift.js" />
+    <None Include="wwwroot\lib\tether\src\js\tether.js" />
+    <None Include="wwwroot\lib\tether\src\js\utils.js" />
   </ItemGroup>
 
-
   <ItemGroup>
-    <Content Remove="Views\Donate\Campaign_2017-08_matching.cshtml" />
-    <Content Remove="Views\Donate\Campaign_2017-12.cshtml" />
-    <Content Remove="Views\Donate\Index_2017-12.cshtml" />
-    <Content Remove="Views\Donate\WeChat_2017-08.cshtml" />
-  </ItemGroup>
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="2.0.0" />
-    <DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" Version="2.0.0" />
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.0" />
-    <DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" Version="2.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\Views.News._Layout.zh-CN.resx" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="iTextSharp" Version="5.5.13" />
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Facebook" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="2.2.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.MicrosoftAccount" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="2.2.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
+    <PackageReference Include="iTextSharp" Version="5.5.13.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.DataAnnotations" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Localization" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Rewrite" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Session" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="2.2.3" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.3" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer.Design" Version="1.1.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.2.3" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.2.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="2.2.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.4" />
-    <PackageReference Include="RestSharp" Version="106.2.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="3.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.3">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.2" />
+    <PackageReference Include="RestSharp" Version="106.11.2" />
     <PackageReference Include="Stripe.net" Version="13.3.1" />
+    <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="DotNetCore.Kit">
+      <HintPath>dlls\DotNetCore.Kit.dll</HintPath>
+    </Reference>
+    <Reference Include="DotNetCore.Kit.Captcha">
+      <HintPath>dlls\DotNetCore.Kit.Captcha.dll</HintPath>
+    </Reference>
+    <Reference Include="DotNetCore.Kit.Serialization">
+      <HintPath>dlls\DotNetCore.Kit.Serialization.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
   <ItemGroup>
     <Compile Update="Resources\DataAnnotations.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
       <DependentUpon>DataAnnotations.resx</DependentUpon>
     </Compile>
-    <Compile Update="Resources\Views.Manage._AddNewCard.zh-CN.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Views.Manage._AddNewCard.zh-CN.resx</DependentUpon>
-    </Compile>
   </ItemGroup>
+
   <ItemGroup>
-    <Content Update="Views\Donate\Campaign_2018a.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="Views\Home\Index_2018a.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="Views\News\2018a.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="Views\News\Index.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="Views\Shared\_ExceptionMail.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt35-webfont.eot">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt35-webfont.svg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt35-webfont.ttf">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt35-webfont.woff">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt35-webfont.woff2">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt65-webfont.eot">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt65-webfont.svg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt65-webfont.ttf">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt65-webfont.woff">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\fonts\dpt65-webfont.woff2">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Update="Resources\Controllers.DonateController.zh-CN.resx">
-      <Generator></Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Controllers.ManageController.zh-CN.resx">
-      <Generator></Generator>
-    </EmbeddedResource>
     <EmbeddedResource Update="Resources\DataAnnotations.resx">
       <Generator>PublicResXFileCodeGenerator</Generator>
       <LastGenOutput>DataAnnotations.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Update="Resources\DataAnnotations.zh-CN.resx">
-      <Generator>PublicResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\ViewModels.CustomerPaymentViewModel.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\ViewModels.PaymentCycle.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.Index.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donate.Campaign_2019b.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.2019a.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donate.Campaign_2019a.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Account.Lockout.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donate.Campaign_2018b.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.2018b.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donate.Campaign_2018a.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.2018a.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Account.ExternalLoginConfirmation.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Account.ResetPasswordConfirmation.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donation.RePayment.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donation.Thanks.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donate.Campaign.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Error.404.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Error.500.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.2017b.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Account.Login.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Donation.Payment.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Manage.Index.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Manage._AddNewCard.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Views.Manage._AddNewCard.zh-CN.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Manage._UpdateCard.zh-cn.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.News.2019b.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Shared.Components.Subscriptions.View.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.Shared._HeadPartial.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Resources\Views.ThankYou.Index.zh-CN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
   </ItemGroup>
-  
+
+
+
 </Project>
