@@ -1,4 +1,9 @@
-﻿@model ExternalLoginConfirmationViewModel
+﻿
+@using Microsoft.AspNetCore.Mvc.Localization
+@using Microsoft.AspNetCore.Identity
+
+@model RenewalWebsite.Models.AccountViewModels.ExternalLoginConfirmationViewModel
+@inject IViewLocalizer Localizer
 @{
     ViewData["Title"] = Localizer["Register"];
 }
@@ -13,7 +18,7 @@
 <div class="form-group">
     <label asp-for="Email" class="col-md-2 control-label">@Localizer["Email"]</label>
     <div class="col-md-10">
-        <input asp-for="Email" class="form-control" disabled="disabled"/>
+        <input asp-for="Email" class="form-control" disabled="disabled" />
     </div>
 </div>
 <div class="form-group">
