@@ -1,12 +1,16 @@
-﻿@model ForgotPasswordViewModel
+﻿
+@using Microsoft.AspNetCore.Mvc.Localization
+@using Microsoft.AspNetCore.Identity
+@inject IViewLocalizer Localizer
+@model RenewalWebsite.Models.AccountViewModels.ForgotPasswordViewModel
 @{
     ViewData["Title"] = @Localizer["Forgot your password?"];
 }
 
 <h2>@ViewData["Title"]</h2>
 @*<p>
-    For more information on how to enable reset password please see this <a href="https://go.microsoft.com/fwlink/?LinkID=532713">article</a>.
-</p>*@
+        For more information on how to enable reset password please see this <a href="https://go.microsoft.com/fwlink/?LinkID=532713">article</a>.
+    </p>*@
 
 <form asp-controller="Account" asp-action="ForgotPassword" method="post" class="form-horizontal">
     <h4>@Localizer["Enter your email."]</h4>
