@@ -1,10 +1,12 @@
 ﻿@using System.Collections.Generic
-@using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
+
+@*@using Microsoft.AspNetCore.Http*@
+@*@using Microsoft.AspNetCore.Authentication*@
 @using Microsoft.AspNetCore.Mvc.Localization
-@model LoginViewModel
+@using RenewalWebsite.Models
+@model RenewalWebsite.Models.AccountViewModels.LoginViewModel
 @inject IViewLocalizer Localizer
-@inject SignInManager<ApplicationUser> SignInManager
+@inject Microsoft.AspNetCore.Identity.SignInManager<ApplicationUser> SignInManager
 
 @{
     ViewData["Title"] = Localizer["Log in"];
