@@ -1,7 +1,9 @@
-﻿@model ResetPasswordViewModel
+﻿@model RenewalWebsite.Models.AccountViewModels.ResetPasswordViewModel
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @{
     ViewData["Title"] = Localizer["Reset password"];
-    
+
 }
 
 <h2>@ViewData["Title"]</h2>
@@ -13,12 +15,12 @@
     <input asp-for="Code" type="hidden" />
     <input asp-for="UserId" type="hidden" />
     @*<div class="form-group">
-        <label asp-for="Email" class="col-md-2 control-label">@Localizer["Email"]</label>
-        <div class="col-md-10">
-            <input asp-for="Email" class="form-control" />
-            <span asp-validation-for="Email" class="text-danger"></span>
-        </div>
-    </div>*@
+            <label asp-for="Email" class="col-md-2 control-label">@Localizer["Email"]</label>
+            <div class="col-md-10">
+                <input asp-for="Email" class="form-control" />
+                <span asp-validation-for="Email" class="text-danger"></span>
+            </div>
+        </div>*@
     <div class="form-group">
         <label asp-for="Password" class="col-md-2 control-label">@Localizer["Password"]</label>
         <div class="col-md-10">
