@@ -1,9 +1,11 @@
-﻿@{
+﻿@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
+@{
     ViewData["Title"] = @Localizer["Reset Password Confirmation"];
-    
+
 }
 
 <h1>@ViewData["Title"]</h1>
 <p>
-   @Localizer["Your password has been reset. Please"] <a asp-controller="Account" asp-action="Login">@Localizer["Click here to log in."]</a>
+    @Localizer["Your password has been reset. Please"] <a asp-controller="Account" asp-action="Login">@Localizer["Click here to log in."]</a>
 </p>
