@@ -1,4 +1,4 @@
-﻿@model SendCodeViewModel
+﻿@model RenewalWebsite.Models.AccountViewModels.SendCodeViewModel
 @{
     ViewData["Title"] = "Send Verification Code";
 }
