@@ -1,4 +1,4 @@
-﻿@model VerifyCodeViewModel
+﻿@model RenewalWebsite.Models.AccountViewModels.VerifyCodeViewModel
 @{
     ViewData["Title"] = "Verify";
 }
