@@ -1,7 +1,7 @@
 ﻿@using System.Globalization
 @using RenewalWebsite.Services
 @using Microsoft.AspNetCore.Mvc.Localization
-@model DonationViewModel
+@model RenewalWebsite.Models.DonationViewModel
 
 @inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
