@@ -1,7 +1,7 @@
 ﻿@using System.Globalization
 @using RenewalWebsite.Services
 @using Microsoft.AspNetCore.Mvc.Localization
-@model DonationViewModel
+@model RenewalWebsite.Models.DonationViewModel
 
 @inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
@@ -47,47 +47,47 @@
             <p>@Localizer["Wanting a job, a family, a safe place to live - these seem like simple and achievable dreams. But realizing these dreams can be almost impossible without the right resources."]</p>
             <p>@Localizer["That's why your support started the residential job program. It gives people an opportunity to build the skills they need to find fulfilling jobs and to thrive in Shanghai (and beyond)."]</p>
             <p>@Localizer["Please help hard-working people gain the skills they need to make a better life."]</p>
-                         @if (currency.Name.Contains("en"))
-                         {
-                    <form asp-controller="Donate" asp-action="CreateCampaign">
+            @if (currency.Name.Contains("en"))
+            {
+                <form asp-controller="Donate" asp-action="CreateCampaign">
 
-                        <div asp-validation-summary="All" class="text-danger"></div>
-                        <div id="textMainMessage" class="text-danger" style="display:none;">
-                            <ul>
-                                <li id="textMainErrorMessage"></li>
-                            </ul>
-                        </div>
-                        <div class="form-group" style="overflow:inherit;">
-                            <div class="input-group">
-                                <label class="mtp-auto">
-                                    @Localizer["I would like to make a"]&nbsp;
-                                </label>
-                                <div class="box">
-                                    <select asp-for="CycleId" class="wide select-font" style="min-width:100px; max-width:100px;border-radius:0.25rem;"
-                                            asp-items="@Model.DonationCycles"></select>
-                                </div>&nbsp;
-                                <label class="mtp-auto">@Localizer["gift of:"]</label>
-                            </div>
-                        </div>
-                        <input type="hidden" id="Reason" name="Reason" />
-                        <input type="hidden" id="IsCustom" name="IsCustom" />
-                        <input type="hidden" id="DonationAmount" name="DonationAmount" />
-                        <ul class="list-unstyled rmb-options mb-5">
-                            <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="1" data-value="false" checked="checked" data-donation-value="28.89" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)28.89</strong> @Localizer["to provide three nights of shelter for a homeless man or woman."]</li>
-                            <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="2" data-value="false" data-donation-value="67.41" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)67.41</strong> @Localizer["to provide one week of shelter and good meals."]</li>
-                            <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="3" data-value="false" data-donation-value="288.9" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)288.90</strong> @Localizer["to provide one month hot meals and care."]</li>
-                            <li>
-                                <div class="input-group">
-                                    <input id="SelectedAmount" name="SelectedAmount" type="radio" value="4" class="rb-align" data-value="true" data-reason="to help as many people as I can." onclick="setFocus(this)" /><strong>@CurrencyService.GetSymbol(currency)</strong>
-                                    <input type="text" id="CustomAmount" name="CustomAmount" onfocus="checkSelection()" class="form-control" style="height:38px;" /> @Localizer["to help as many people as I can."]
-                                </div>
-                            </li>
+                    <div asp-validation-summary="All" class="text-danger"></div>
+                    <div id="textMainMessage" class="text-danger" style="display:none;">
+                        <ul>
+                            <li id="textMainErrorMessage"></li>
                         </ul>
-                        <div class="clearfix mt-3 mb-3">
-                            <button id="btnSave" class="btn btn-secondary" type="button" onclick="checkAmount();">@Localizer["DONATE WITH CREDIT OR DEBIT CARD NOW"]&nbsp;<i class="arrow right-arrow"></i></button>
-                            <button id="btnSaveFinal" class="btn btn-secondary hide" type="submit">@Localizer["DONATE WITH CREDIT OR DEBIT CARD NOW"]&nbsp;<i class="arrow right-arrow"></i></button>
+                    </div>
+                    <div class="form-group" style="overflow:inherit;">
+                        <div class="input-group">
+                            <label class="mtp-auto">
+                                @Localizer["I would like to make a"]&nbsp;
+                            </label>
+                            <div class="box">
+                                <select asp-for="CycleId" class="wide select-font" style="min-width:100px; max-width:100px;border-radius:0.25rem;"
+                                        asp-items="@Model.DonationCycles"></select>
+                            </div>&nbsp;
+                            <label class="mtp-auto">@Localizer["gift of:"]</label>
                         </div>
-<p>@await Html.PartialAsync("_CurrencyTogglePartial")</p>
+                    </div>
+                    <input type="hidden" id="Reason" name="Reason" />
+                    <input type="hidden" id="IsCustom" name="IsCustom" />
+                    <input type="hidden" id="DonationAmount" name="DonationAmount" />
+                    <ul class="list-unstyled rmb-options mb-5">
+                        <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="1" data-value="false" checked="checked" data-donation-value="28.89" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)28.89</strong> @Localizer["to provide three nights of shelter for a homeless man or woman."]</li>
+                        <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="2" data-value="false" data-donation-value="67.41" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)67.41</strong> @Localizer["to provide one week of shelter and good meals."]</li>
+                        <li> <input id="SelectedAmount" name="SelectedAmount" type="radio" value="3" data-value="false" data-donation-value="288.9" data-reason="For most urgent needs" /><strong>@CurrencyService.GetSymbol(currency)288.90</strong> @Localizer["to provide one month hot meals and care."]</li>
+                        <li>
+                            <div class="input-group">
+                                <input id="SelectedAmount" name="SelectedAmount" type="radio" value="4" class="rb-align" data-value="true" data-reason="to help as many people as I can." onclick="setFocus(this)" /><strong>@CurrencyService.GetSymbol(currency)</strong>
+                                <input type="text" id="CustomAmount" name="CustomAmount" onfocus="checkSelection()" class="form-control" style="height:38px;" /> @Localizer["to help as many people as I can."]
+                            </div>
+                        </li>
+                    </ul>
+                    <div class="clearfix mt-3 mb-3">
+                        <button id="btnSave" class="btn btn-secondary" type="button" onclick="checkAmount();">@Localizer["DONATE WITH CREDIT OR DEBIT CARD NOW"]&nbsp;<i class="arrow right-arrow"></i></button>
+                        <button id="btnSaveFinal" class="btn btn-secondary hide" type="submit">@Localizer["DONATE WITH CREDIT OR DEBIT CARD NOW"]&nbsp;<i class="arrow right-arrow"></i></button>
+                    </div>
+                    <p>@await Html.PartialAsync("_CurrencyTogglePartial")</p>
                 </form>
             }
             else
