@@ -1,7 +1,7 @@
 ﻿@using System.Globalization
 @using RenewalWebsite.Services
 @using Microsoft.AspNetCore.Mvc.Localization
-@model DonationViewModel
+@model RenewalWebsite.Models.DonationViewModel
 
 @inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
@@ -92,7 +92,7 @@
             <div class="clearfix mt-3 mb-3">
                 <a class="btn btn-secondary" target="_blank" href="https://yoopay.cn/pay/renewal365">@Localizer["DONATE NOW USING WECHAT OR ALIPAY"] <i class="arrow right-arrow"></i></a>
             </div>
- 
+
             <p class="m-tb-15">@await Html.PartialAsync("_CurrencyTogglePartial")</p>
         }
     </div>
