@@ -1,7 +1,8 @@
 ﻿@{
     ViewData["Title"] = Localizer["Downloads"];
 }
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 <div class="row">
     <div class="col-lg-6 offset-lg-0 offset-md-2 col-md-8 offset-sm-1">
         <img src="~/images/home/slide-ziggy.jpg" class="figure-img img-fluid" alt="download page image">
