@@ -1,5 +1,6 @@
 ﻿@using RenewalWebsite.Services
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
 
 @{
