@@ -1,5 +1,6 @@
 ﻿@using RenewalWebsite.Services
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
 
 @{
@@ -27,12 +28,13 @@
         <p>@Localizer["How do you make a new life when everything you have has been taken away? It’s only possible with a helping hand! That’s why the people in our residential job training program need you."]</p>
         <p>@Localizer["The homeless people who come to the Renewal Center want to work, but they don’t have job skills or even a place to live."]
         <p>@Localizer["With your help, we can welcome people into a long-term program that gives them all of the tools they need to live independent and fulfilling lives. Please send a gift toward this life-changing program."]
-        <p>@Localizer[""]
+        <p>
+            @Localizer[""]
 
         @*<p><a asp-controller="Donate" asp-action="Campaign">@Localizer["My gift to help people start a new, fulfilling life!"]</a></p>*@
 
         <div class="clearfix mt-4 mb-5">
             <a class="btn btn-secondary btn-block" asp-controller="Donate" asp-action="Campaign">@Localizer["My gift to help people start a new, fulfilling life!"] <i class="arrow right-arrow"></i></a>
         </div>
-    </div>
+</div>
 </div>
