@@ -1,4 +1,5 @@
-﻿@model AddPhoneNumberViewModel
+﻿
+@model RenewalWebsite.Models.ManageViewModels.AddPhoneNumberViewModel
 @{
     ViewData["Title"] = "Add Phone Number";
 }
