@@ -1,4 +1,6 @@
-﻿@model ChangePasswordViewModel
+﻿@model RenewalWebsite.Models.ManageViewModels.ChangePasswordViewModel
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @{
     ViewData["Title"] = Localizer["Change Password"];
 }
