@@ -1,8 +1,9 @@
-﻿@model SetPasswordViewModel
+﻿@model RenewalWebsite.Models.ManageViewModels.SetPasswordViewModel
 @{
     ViewData["Title"] = "Set Password";
 }
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 <p class="text-gray-dark">
     @Localizer["Set a local password so you can log in without an external provider."]
 </p>
