@@ -1,4 +1,4 @@
-﻿@model VerifyPhoneNumberViewModel
+﻿@model RenewalWebsite.Models.ManageViewModels.VerifyPhoneNumberViewModel
 @{
     ViewData["Title"] = "Verify Phone Number";
 }
