@@ -3,7 +3,7 @@
 @using Microsoft.AspNetCore.Localization
 @using Microsoft.AspNetCore.Mvc.Localization
 @using Microsoft.AspNetCore.Mvc.Rendering
-
+@inject IViewLocalizer Localizer
 @inject ICurrencyService CurrencyService
 
 @{
