@@ -1,4 +1,4 @@
-﻿@model EventLog
+﻿@model RenewalWebsite.Models.EventLog
 
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml">
