@@ -1,7 +1,8 @@
 ﻿@using System.Globalization
 @using Microsoft.AspNetCore.Identity
 @using RenewalWebsite.Models
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
 
@@ -49,8 +50,8 @@
                         <a class="nav-link" asp-controller="Donate" asp-action="Index">@Localizer["Donate"]</a>
                     </li>
                     @*<li class="nav-item">
-                        <a class="nav-link" asp-controller="Downloads" asp-action="Index">@Localizer["Download"]</a>
-                    </li>*@
+                            <a class="nav-link" asp-controller="Downloads" asp-action="Index">@Localizer["Download"]</a>
+                        </li>*@
                 </ul>
                 @await Html.PartialAsync("_LoginPartial")
             </div>
