@@ -1,6 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using RenewalWebsite.Models
-
+@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
 
