@@ -1,4 +1,5 @@
-﻿
+﻿@using Microsoft.AspNetCore.Mvc.Localization
+@inject IViewLocalizer Localizer
 @{
     ViewData["Title"] = Localizer["Subscribe"];
     Layout = "~/Views/Shared/_Layout.cshtml";
