@@ -1,9 +1,3 @@
-﻿@using RenewalWebsite
-@using RenewalWebsite.Models
-@using RenewalWebsite.Models.AccountViewModels
-@using RenewalWebsite.Models.ManageViewModels
-@using Microsoft.AspNetCore.Identity
-@using Microsoft.AspNetCore.Mvc.Localization
+﻿@using TestWebsite
+@using TestWebsite.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-
-@inject IViewLocalizer Localizer
